requirements.txt
fastapi>=0.115.0
pydantic-settings>=2.3.0
email-validator>=2.1.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pillow>=10.4.0
httpx==0.27.0
python-dotenv==1.0.1
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
tensorflow>=2.16.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
google-generativeai>=0.7.0
